\name{plotspec}
\alias{plotspec}
\title{
  Plot spectrum output from \code{\link{spec}}
}
\description{
  Plot spectrum output from \code{\link{spec}}
}
\usage{
plotspec(x, fit, np = TRUE, ...)
}
\arguments{
  \item{x}{
    Wavelength values
}
  \item{fit}{
    Output from \code{\link{spec}}
}
\item{np}{
  If TRUE, new plot is generated.
}
\item{...}{
  Other plotting options
}
}
\references{
Raymond K. W. Wong, Vinay L. Kashyap, Thomas C. M. Lee and David A. van Dyk (2015). \emph{Detecting Abrupt Changes in the Spectra of High-energy Astrophysical Sources}. Submitted.
}

\seealso{
\code{\link{spec}}
}
